+++ Triangulation_2/include/CGAL/draw_triangulation_2.h
@@ -17,6 +17,7 @@
 
 #ifdef CGAL_USE_BASIC_VIEWER
 
+#include <CGAL/Qt/init_ogl_context.h>
 #include <CGAL/Triangulation_2.h>
 #include <CGAL/Random.h>
 
@@ -144,6 +145,7 @@ void draw(const CGAL_T2_TYPE& at2,
 
   if (!cgal_test_suite)
   {
+    CGAL::Qt::init_ogl_context(4,3);
     int argc=1;
     const char* argv[2]={"t2_viewer","\0"};
     QApplication app(argc,const_cast<char**>(argv));