+++ Installation/CHANGES.md
@@ -25,11 +25,14 @@ A comprehensive list of the supported file formats is available in the Stream_su
 -   Added a filtering mechanism so that costly tests get only applied to the next candidate for the edge collapse.
 -   Added the class `Polyhedral_envelope_filter` that enables to perform mesh simplification inside a polyhedral envelope of the input mesh.
 
+### 2D Regularized Boolean Set Operations
+-   Added documentation for the free functions `oriented_side(const Point_2& p, ....)` that accept a point and a polygon, and improved the documentation of all other functions
+
 ### [2D and 3D Alpha Shapes](https://doc.cgal.org/5.3/Manual/packages.html#PkgAlphaShapes2)
 -   The following deprecated classes have been removed: `Alpha_shape_euclidean_traits_2`,
     `Weighted_alpha_shape_euclidean_traits_2`, `Alpha_shape_euclidean_traits_3`, and
     `Weighted_alpha_shape_euclidean_traits_3`. All CGAL kernel can be used directly as models
-    of the Alpha Shape concepts.
+    of the concepts of the 2D and 3D Alpha Shape packages.
 
 [Release 5.2](https://github.com/CGAL/cgal/releases/tag/v5.2)
 -----------